Remove pointers from proposal
This strips the proposal down to just reference types, pointers can be addressed separately. This also strips out the deprecation of `in`, `out` and `inout` parameter annotations to be discussed seprately. Lastly I added a section at the end to list oustanding questions from the review that need to be addressed in the detailed design.