Initial HLSL Specifiction LaTeX stub (#10)
This stubs out some LaTeX goop and a first stab at the Introduction for
an HLSL Specification.

I'm sure LaTeX isn't everyone's preferred format, but it is git-friendly
and supports all the formatting bells and whistles we might need.

This initial commit adds a glossary as well as formatting for chapter, section
and subsection headings which include stable section labels. The ISO C++
standard, added stable section labels so that as chapter and section numbers
change there are stable references that persist.