Add GitHub pages build
This change adds support for builing a GitHub pages static site reproducing the content of the GitHub markdown as well as built versions of the LaTeX specification. This change is a starting point, there is still a lot of work to improve the visual look and functionality of the site as well as the LaTeX to HTML generation.

A preview of the site is available off my fork of the repository here:
https://llvm-beanz.github.io/hlsl-specs

Note: Some links that are hard-coded in the markdown files will not work until the page is generated and deployed to GitHub Pages.