Change how raw regions are marked
This fixes rendering issues in 0002 and 0005. It also puts the raw
marker into an HTML comment so that it doesn't display in GitHub's
markdown preview. I've added the raw markings to the template too so
future proposals will inherit it for the full proposal.

I suspect using liquid markup in proposals will be rare if ever.