Add GitHub pages build (#27)
This change adds support for builing a GitHub pages static site reproducing the content of the GitHub markdown as well as built versions of the LaTeX specification. This change is a starting point, there is still a lot of work to improve the visual look and functionality of the site as well as the LaTeX to HTML generation.

A preview of the site is available off my fork of the repository here:
https://llvm-beanz.github.io/hlsl-specs

Note: Some links that are hard-coded in the markdown files will not work until the page is generated and deployed to GitHub Pages.

* Change how raw regions are marked

This fixes rendering issues in 0002 and 0005. It also puts the raw
marker into an HTML comment so that it doesn't display in GitHub's
markdown preview. I've added the raw markings to the template too so
future proposals will inherit it for the full proposal.

I suspect using liquid markup in proposals will be rare if ever.