Add stub grammar definitions and sections
These are just empty placeholder sections with some grammar definitions
to fill in the ID expression grammars. Properly documenting this will
require going back and defining scopes and other lexical constructs.

Since all of this is more or less inherited from C/C++ I'm going to
leave them as stubs for now and focus on the HLSL-specific sections.