Add design considerations and clarify process (#46)
This update adds new documents on design considerations and contact information.

The design considerations document tries to codify some of the basic design language for HLSL and cover common topics that features will need to address in their specification.

This also updates some wording and details around the process docs and specifically about proposal sponsors. We think it is important that Vulkan-specific proposals have a sponsor from the Vulkan Working Group, so that is explicitly stated.