[0010] Allow casting to uint64_t (#134)
* [0010] Allow casting to uint64_t

We have decided to allow casting to uint64_t. It will make create
more opportunities for developers to make mistakes, but they do not
have to use it. The implementation cost does not seem too high.

Fixes https://github.com/microsoft/hlsl-specs/issues/93

* Remove cast to bool

As Greg suggested, the cast to bool is no longer needed if we can cast
to an int, so we removed it. This commit also adds the uint64_t cast to
the pseudo class definition for BufferPointers.

* Fix linked list example after rebaseing.