Update conversion sequences and describe ranks
This change fixes a spec bug in the description of HLSL truncation
casts, restricts the array-to-pointer cast to unbounded arrays (where
they decay to pointers), and specifies integral and floating point
conversion ranks.

This updates the spec to be more consistent with how the language is
being implemented in clang and lays the groundwork for describing
overload resolution rules.