Create shader model spec template (#148)
The existing basic template has a lot of useful information, but lacks
some things that are of interest when documenting shader model features.
This alters the format of the versioning, adds details about language,
interchange format, diagnostics, and runtime/driver interfaces to the
implementation details section including questions to prompt details
that might be important to document. In addition, testing and migration
sections are included to go into how the feature can best be vetted and
adopted.