reformat and clarify a few things
reformat to comply with md best practices. Demoting the attributes from
headers wasn't strictly necessary, but it solved a problem and I think
the appearance is better and has less weird characters in what is
ostensibly a linkable header.

Reclarify a few things resulting from reviewing older suggestions and
reformatting