[Language] Begin Basic.types (#198)
* [Language] Begin Basic.types

This begins drafting the Basic.types section which specifies the
foundations of HLSL's data types. This includes some basic information
about representations as well as the classifications of types and the
basic behaviors of some types.

I'm coining a new term "intangible types" to describe a type that
is has no known size for an object or value representation, but behaves
as a complete type in local memory.

The definition of "intangible types" will play a larger role in some of
the wider compiler behaviors as we work through things like call
behavior.