[0017] Add proposal for mitigations (#187)
* Add proposal for mitigations

Mitigations for the impact of this change will be the introduction of
new diagnostics and reliance on existing diagnostics.

* Add notes about integer behavior and min precision types

This captures Tex's comments about integer behavior and tries to clarify
documentation for minimum precision types. It also adds notes to the
language specification PRs that document the proposed 202x behaviors.