Add cbuffer and tbuffer to examples
cbuffer and tbuffer are odd. They declare global variables that are
read from a specific cbv or srv and allow defining layouts of the
buffers. The name of the buffer is exposed in reflection data, but not
declared as a name within the translation unit.

The names of variables within the buffers are declared as global names.