Updates to process docs (#219)
This clarifies details for the process docs for how proposals move
through phases. This update also makes some small tweaks to the process
for infrastructure proposals to optionally bypass the **Under Review**
phase when outside feedback is not expected. This allows for a more
expedited process for designs that only impact internal implementation
details.

This also contains a few small updates to the infrastructure tempalte
to have an optional link to LLVM's Discourse, and more clarification
around documentation expectations.