Add `export` keyword to HLSL language spec (#268)
Add export keyword to HLSL language specs. The spec describes two ways the export keyword can be used. One way is to use it on individual function declarations like `export void f() {}`. The other way is to use it on a group of function declarations:

export {
   void f1();
   void f2() {}
}

DXC currently supports only the first case, but since Clang has support for both, we might as well support it in HLSL 202y too.

This spec update does not yet include detailed rules for when a function can or cannot be exported, such as when it has resource argument or semantic annotations. That will be covered by llvm/llvm-project#93330.

This change also adds more sections under Declarations (based on C++ spec layout).

Contributes to: llvm/llvm-project#92812