[0011] Allows cap and ext attributes on variable declarations (#270)
To be able to add the necessary capabilities and extensions for
SpirvTypes, we have to be able to add the capabilities and extension
attributes to the variables of that type. This change updates the inline
spirv spec to allow adding the capabilities and extension attributes to
type aliases, variables, and fields.