[INF-0004] New proposal for validator hashing
This proposal seeks to address user needs for building DirectX shaders
on a wider variety of platforms than we can reasonably support through
official releases, as well as the need for certain workflows to bypass
the DXIL validator.

This proposal includes a plan for releasing the validator hashing
algorithm into the DXC public repository, and includes the algorithms in
the appendix.

Strategically this aligns the DXIL validator to fill a role more aligned
with other virtual ISA validation tools.