Fix up load and store functions
We add the ability to do loads and stores from groupshared memory. This
involved adding a WorkgroupSpirvPointer type that is used by the Load
and Store functions. We also create a GetGroupSharedAddress function to
get the address.

The other load and store function are refactored to make the layout a
parameter, to move the memory access mask to a template parameter, and
to always have a stride. All memory layouts require an explicit stride,
so even though the stride is optional on the instructions, there is not
case where it can be ommitted.