Clarifications and alignment of vector truncations
Previously the spec had allowed trunction conversions to produce
cxvalues which enabled them to be used in output arguments. This is
inconsistent with the DXC and FXC implementation, which disallows
binding trunctions to output arguments.

The spec also previously did not allow trunction to scalar values.
Initially that was excluded due to discussions around truncations and
extensions of output arguments. Since DXC and FXC both disallow
trunction in output arguments, it seems safe to support truncation to
scalar here with a cooresponding warning.

This will allow the spec to more consistently align with the reference
implementations.

This has no change to overload resolution except that a vector argument
can be truncated to a scalar producing a Trunctaion, Truncation
Prootion, or Truncation Conversion implicit conversion sequence.: