[0007] Big update to flesh out const instance methods (#34)
* Big update to flesh out const instance methods

I started digging a bit into this in DXC and identified a lot of cross-cutting issues that we need to consider in this.

This change updates the proposal to flesh out the behavior and related issues around HLSL const-correctness.

This does identify some source-breaking changes that this would introduce as well as some behavior changes that I think we should introduce in the HLSL built-in types.

* Add raw liquid blocks

* Updates based on feedback from @tex3d

* instance-methods -> non-static member functions

* Move to 202y