[0002] Specify the grammar formulations for attributes (#65)
* Specify the grammar formulations for attributes

This does not exhaustively express all the places C++ attributes are
valid, but it should cover enough detail to be useful.

Fixes #63.

* Updates based on feedback from @bogner

* Updating grammar specifiction

This updates the grammar specification to align with how C++ specifies
parsing grammar, and the language spec latex is included. I've also
added images containing the rendered grammar for ease of reading.

* Fix image paths

* Update proposals/0002-cxx-attributes.md

Co-authored-by: Ashley Coleman <ascoleman@microsoft.com>

---------

Co-authored-by: Ashley Coleman <ascoleman@microsoft.com>