Add an initial proposal for HLSL 202x deprecations
This proposal seeks to enumerate the smaller features that should be
deprecated or significantly tightened in HLSL 202x, which are not
large enough to indipendently warrant their own proposal documents.

These features represent technical debt in DXC's implementation of HLSL
which should be removed in future language versions to avoid carrying it
forward indefinitely.