[202x] Propose adding vk::SampledTexture* types (#343)
* [202x] Propose adding vk::SampledTexture* types

This proposal adds a set of `vk::SampledTexture*` types to represent
SPIR-V's sampled image types. The intent is to replace and deprecate the
`[[vk::combinedImageSampler]]` annotation.

I have created this with the extension template with the intent of using
the fast-track process, as this is a platform-specific feature that is
already supported by all Vulkan hardware and drivers, and it does not
change any existing language behavior.

* Refinements

* Mention portability

* Address deprecation and feature check

* Number the proposal