Updating and extending the linalg::Matrix APIs
The two main changes here are adding the matrix use and matrix scope
template arguments.

The matrix use distinguishes which argument a matrix can be used as in
matrix operations. A matrix can be either the left hand operand (A), the
right hand operand (B), or a result (accumulator).

The matrix scope denotes the scope of uniformity of matrix instances.
All matrix-matrix operations in the current proposed header require
wave-scope matrices, however vector-matrix operations may operate on
either wave-scope or thread-scope matrices.