[0035] `linalg::Matrix` detailed design revision (#556)
This is a big revision to the `linalg::Matrix` proposal. There are a few
substantial changes in this proposal to the API formulation.

First it updates the `Matrix` template so that the component type is
specified as an enumeration rather than an explicit type. This allows
specifying matrix types that are not HLSL built-in types.

Second, it adds two new arguments to the `Matrix` template: `Use` and
`Scope`. The `MatrixUse` enumeration specifies the type of matrix as an
`A`, `B`, or `Accumulator`. This aligns with the Vulkan
cooperative_matrix naming rather than the earlier DirectX `Left` and
`Right` naming in an effort to be less confusing to users who are
supporting both platforms. The additional `MatrixScope` enumeration
denotes the scope of uniformity that the matrix value is operating
under. Wave-uniformity is required for matrix-matrix operations, however
vector-matrix operations can operate on thread-divergent matrices.

This proposal adds support for vector-matrix operations subsuming the
features described in [cooperative
vectors](https://github.com/microsoft/hlsl-specs/blob/main/proposals/0029-cooperative-vector.md).

Last, this proposal adds some rough initial DXIL operation
specifications and updates the list of outstanding questions.