[0035] Add Multiply overload for explicit output type
This adds a second overload of linalg::Multiply to support differing
input matrix types if an explicit output type is specified. To make sure
this works I also fixed the SFINAE magic in the HLSL implementation
which caused some cascading source updates.

This additional overlod was mentioned in the feedback in #257