[0035] Revise matrix element accesses
This PR removes the row and element accessors in favor of a more general
means of accessing per-thread vectors. This change allows initializing a
matrix from per-thread vectors as well as splitting it out into
constituent vectors.

The vectors represent rows in A matrices, and columns in B matrices.
Since the layout of Accumulator matrices vary by vendor, this method may
only be used to construct A or B matrices.

Resolves #575