Restrict `MatrixScope::Thread` to be read-only (#625)
This PR restricts the usage of thread scope matrices such that they can
be considered read-only and never have to be fully materialized in
registers. The matrix can only loaded from a ByteAddressBuffer and used
as a source of vector-matrix multiplication. This will allow some
implementations to fold the load and matmulvec operations. This change
better aligns this specification with the cooperative vector usage
models.

Note: another PR will be submitted to address OuterProductAccumulate for
thread scope matrices.