[0041] TriangleObjectPositions proposal (#632)
Replaces #571 and #572.

I've also fleshed out the doc slightly, to codify that this is targeting
shader model 6.10, and now that shader model 6.9 is available, to
unresolve the DXIL opcode discussion about return type to mention that
`float3` and `vector<float, 9>` are available as options. I added a
SPIR-V section and references to DXIL SER as well.

---------

Co-authored-by: Damyan Pepper <damyanp@microsoft.com>