[SM6.9] Propose making some optional features required in SM 6.9 (#642)
Shader Model 6.9 introduces vectorized DXIL which means that we need to add new conformance tests for all DXIL operations that now accept vectors. Some of these operations are optional, and so we'd need to test combinations of these features.

In practice every device that supports SM 6.9 is expected to support these operations, so we can simplify the testing matrix by assuming that these operations are supported.