[Lang] Update and restructure cbuffer spec
This change pulls the cbuffer section into the Declarations chapter
instead of the Resources chapter since it defines a declaration type,
and updates the grammar and specification language based on the updates
to proposal 37.

The updated language avoids nesting of cbuffers and namepaces in
cbuffers to simplify semantic validation. The updated langauge also adds
additional language around storage durations which is referenced in the
updated specification language.