[0035] Add Cooperative Vector interpretations
This adds in the bias and input vector interpretation arguments to the
linalg vector-matrix operations. To support these new arguments coming
in from HLSL the API gains a new dx::linalg::InterpretedVector object,
and a related dx::linalg::VectorRef object.

The dx::linalg::VectorRef object can be used to represent an
InterpretedVector stored in a ByteAddressBuffer and can be used to
materialize the vector into local memory immediately before the matmul
call.