[0037] Only restrict nesting cbuffer and namespace decls (#685)
This slightly scales back the proposal to restricting namespaces in
cbuffers and restricting nesting of cbuffers. This is likely a safe
change for HLSL 202x, and we'll consider more significant changes in
subsequent language updates.

This conclusion came out of the language design meeting discussion from
10/13/2025.

The language spec updates are posted in a draft as
https://github.com/microsoft/hlsl-specs/pull/684.

---------

Co-authored-by: Helena Kotas <hekotas@microsoft.com>