Revise createMatrix to create a unique reference
After losing all its type information to annotateMatrix, createMatrix had no purpose and would have resulted in all calls being merged as redundant, however, some means of uniquely identifying a matrix is needed that annotateMatrix can't do because it needs to be generated for parameters as well as locals and globals (assuming we allow globals).

Previous handles for resources have had global variables that are able to uniquely identify them and workgraph handles have had parameters to the entry functions. Lacking either of these, it was impossible to distinguish allocas created for the initial declaration of a matrix and one that was created for a parameter that should get merged when inlining that function. By generating a unique createMatrix call where the local variable is declared or in the entry block for global variables.

Addresses some issues in #696 though not in the way I initially proposed there