[0035] Align matrix-vector APIs with coopvec (#741)
This aligns the matrix-vector HLSL APIs with the SM 6.9 cooperative
vector feature such that the matrix is an `A` matrix and the vectors are
column vectors rather than row vectors. It also aligns the argument
orders between the HLSL and DXIL APIs to make it easier to read (SM 6.9
had a mismatch between HLSL APIs and DXIL).

---------

Co-authored-by: Michael Bedy <mjb@bedy.org>
Co-authored-by: Ashley Coleman <ascoleman@microsoft.com>